Create an extension for StubbedDataCollection+LaunchEnvironmentEncodable
